+++ src/System.Windows.Forms/src/System/Windows/Forms/ApplicableToButtonAttribute.cs
@@ -1,0 +1,14 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    [AttributeUsage(AttributeTargets.All)]
+    internal sealed class ApplicableToButtonAttribute : Attribute
+    {
+        public ApplicableToButtonAttribute()
+        {
+        }
+    }
+}
