+++ src/System.Windows.Forms/src/System/Windows/Forms/Internal/Gdi/FontCache.FontCacheEntry.cs
@@ -1,0 +1,18 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Drawing;
+using static Interop;
+
+namespace System.Windows.Forms
+{
+    internal sealed partial class FontCache
+    {
+        private sealed class FontCacheEntry : CacheEntry
+        {
+            public FontCacheEntry(Data data, bool tracked) : base(data, tracked) { }
+            public override Gdi32.HFONT Object => Data.HFONT;
+        }
+    }
+}
